refactor: ReservationController로 예약 페이지 요청 처리 위치 변경
