+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -1,6 +1,11 @@
 package roomescape.controller;
 
+import static org.hamcrest.Matchers.is;
+
 import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import java.util.HashMap;
+import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -11,20 +16,67 @@ import org.springframework.test.annotation.DirtiesContext;
 public class ReservationControllerTest {
 
     @Test
-    @DisplayName("/admin으로 요청시 메인 페이지 응답")
-    void getMainPage() {
+    @DisplayName("/reservations 요청 시 예약 정보 조회")
+    void readReservations() {
         RestAssured.given().log().all()
-            .when().get("/admin")
+            .when().get("/reservations")
             .then().log().all()
-            .statusCode(200);
+            .statusCode(200)
+            .body("size()", is(0));
     }
 
     @Test
-    @DisplayName("/admin/reservation 요청 시 예약 관리 페이지 응답")
-    void readReservation() {
+    @DisplayName("예약 관리 페이지 내에서 예약 추가")
+    void createReservation() {
         RestAssured.given().log().all()
-            .when().get("/admin/reservation")
+            .contentType(ContentType.JSON)
+            .body(getTestParamsWithReservation())
+            .when().post("/reservations")
+            .then().log().all()
+            .statusCode(200)
+            .body("name", is("브라운"));
+
+        RestAssured.given().log().all()
+            .when().get("/reservations")
+            .then().log().all()
+            .statusCode(200)
+            .body("size()", is(1));
+    }
+
+    @Test
+    @DisplayName("예약 관리 페이지 내에서 예약 삭제")
+    void deleteReservation() {
+        RestAssured.given().log().all()
+            .contentType(ContentType.JSON)
+            .body(getTestParamsWithReservation())
+            .when().post("/reservations")
+            .then().log().all()
+            .statusCode(200)
+            .body("id", is(1));
+
+        RestAssured.given().log().all()
+            .when().get("/reservations")
+            .then().log().all()
+            .statusCode(200)
+            .body("size()", is(1));
+
+        RestAssured.given().log().all()
+            .when().delete("/reservations/1")
             .then().log().all()
             .statusCode(200);
+
+        RestAssured.given().log().all()
+            .when().get("/reservations")
+            .then().log().all()
+            .statusCode(200)
+            .body("size()", is(0));
+    }
+
+    private Map<String, String> getTestParamsWithReservation() {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("time", "15:40");
+        return params;
     }
 }
